# swag_docs/swagger.yaml
definitions:
  AdminAccountJoined:
    properties:
      created_at:
        type: string
      created_by_name:
        type: string
      created_by_urn:
        type: string
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      email_verified_at_ts:
        type: integer
      external_id:
        type: string
      first_name:
        type: string
      hashed_password:
        type: string
      id:
        type: string
      is_super_user_session:
        type: integer
      last_login_ts:
        type: integer
      last_name:
        type: string
      name:
        type: string
      organization_id:
        type: string
      organization_name:
        type: string
      password_updated_at_ts:
        type: integer
      phone:
        type: string
      properties:
        $ref: '#/definitions/account.Properties'
      role:
        type: integer
      signup_properties:
        $ref: '#/definitions/account.SignupProperties'
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        type: string
      updated_by_name:
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
    type: object
  account.Account:
    properties:
      ChangeLogs:
        type: boolean
      Class:
        type: string
      Client:
        $ref: '#/definitions/Client'
      CreatedAtColumn:
        type: string
      IDColumn:
        type: string
      IsVersioned:
        type: boolean
      ManualCache:
        type: boolean
      PackageName:
        type: string
      SavingUser:
        $ref: '#/definitions/github.com/griffnb/core/lib/model/coremodel.Model'
      Standardized:
        type: boolean
      Table:
        type: string
      UpdatedAtColumn:
        type: string
      created_at:
        format: date-time
        type: string
      created_by_urn:
        type: string
      data:
        additionalProperties: {}
        type: object
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      email_verified_at_ts:
        type: integer
      external_id:
        type: string
      fieldMap:
        additionalProperties: {}
        type: object
      first_name:
        type: string
      hashed_password:
        type: string
      id:
        format: uuid
        type: string
      is_super_user_session:
        type: integer
      isLoaded:
        type: boolean
      isSaved:
        type: boolean
      last_login_ts:
        type: integer
      last_name:
        type: string
      newData:
        additionalProperties: {}
        type: object
      organization_id:
        format: uuid
        type: string
      password_updated_at_ts:
        type: integer
      phone:
        type: string
      preValues:
        additionalProperties: {}
        type: object
      properties:
        $ref: '#/definitions/account.Properties'
      readerSem:
        format: int32
        type: integer
      role:
        type: integer
      saveData:
        additionalProperties: {}
        type: object
      signup_properties:
        $ref: '#/definitions/account.SignupProperties'
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        format: date-time
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
      v:
        format: int32
        type: integer
      writerSem:
        format: int32
        type: integer
    required:
    - id
    - urn
    - created_by_urn
    - updated_by_urn
    - status
    - deleted
    - disabled
    - updated_at
    - created_at
    - first_name
    - last_name
    - email
    - phone
    - external_id
    - test_user_type
    - organization_id
    - role
    - properties
    - signup_properties
    - hashed_password
    - password_updated_at_ts
    - email_verified_at_ts
    - last_login_ts
    - is_super_user_session
    type: object
...
  account.AccountJoinedPublic:
    properties:
      created_at:
        format: date-time
        type: string
      created_by_urn:
        type: string
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      external_id:
        type: string
      first_name:
        type: string
      id:
        format: uuid
        type: string
      is_super_user_session:
        type: integer
      last_name:
        type: string
      organization_id:
        format: uuid
        type: string
      phone:
        type: string
      role:
        type: integer
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        format: date-time
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
    required:
    - id
    - urn
    - created_by_urn
    - updated_by_urn
    - status
    - deleted
    - disabled
    - updated_at
    - created_at
    - first_name
    - last_name
    - email
    - phone
    - external_id
    - test_user_type
    - organization_id
    - role
    - is_super_user_session
    type: object
  account.AccountPublic:
    properties:
      created_at:
        format: date-time
        type: string
      created_by_urn:
        type: string
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      external_id:
        type: string
      first_name:
        type: string
      id:
        format: uuid
        type: string
      is_super_user_session:
        type: integer
      last_name:
        type: string
      organization_id:
        format: uuid
        type: string
      phone:
        type: string
      role:
        type: integer
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        format: date-time
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
    required:
    - id
    - urn
    - created_by_urn
    - updated_by_urn
    - status
    - deleted
    - disabled
    - updated_at
    - created_at
    - first_name
    - last_name
    - email
    - phone
    - external_id
    - test_user_type
    - organization_id
    - role
    - is_super_user_session
    type: object
...
  account.Properties:
    properties:
      external_user_info: {}
      invite_key:
        type: string
      invite_ts:
        format: int64
        type: integer
      last_seen:
        format: int64
        type: integer
      verify_email_key:
        type: string
    type: object
...
  account.SignupProperties:
    properties:
      is_oauth:
        format: int64
        type: integer
    required:
    - is_oauth
    type: object
...
  account_service.TestUserInput:
    properties:
      organization_id:
        type: string
    type: object
...
  accounts.PasswordInput:
    properties:
      current_password:
        type: string
      password:
        type: string
      password_confirmation:
        type: string
    type: object
...
  accounts.ResendVerifyEmailPayload:
    properties:
      cf_token:
        type: string
    type: object
...
  accounts.SetPasswordInput:
    properties:
      password:
        type: string
      password_confirmation:
        type: string
      verify:
        type: string
    type: object
...
  accounts.UpdatePrimaryEmailAddressPayload:
    properties:
      cf_token:
        type: string
      email:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /account/{id}:
    delete:
      consumes:
      - application/json
      description: Marks an account as deleted by setting status to USER_DELETED
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/AdminAccountJoined'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete account
      tags:
      - Account
    get:
      consumes:
      - application/json
      description: Get Account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountJoinedPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Account
      tags:
      - Account
    put:
      consumes:
      - application/json
      description: Update Account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.AccountPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update Account
      tags:
      - Account
...
  /admin/testUser:
    post:
      consumes:
      - application/json
      description: Creates a test account with the provided details and optionally
        logs in as that account
      parameters:
      - description: Test user details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/account_service.TestUserInput'
      - description: Set to login as the created user
        in: query
        name: login_as
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.Account'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create test account
      tags:
      - Account
  /auth/email:
    put:
      consumes:
      - application/json
      description: Updates the user's primary email address for unverified accounts
      parameters:
      - description: Email update details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.UpdatePrimaryEmailAddressPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update primary email
      tags:
      - Account
  /auth/password:
    put:
      consumes:
      - application/json
      description: Updates the user's password after verifying current password
      parameters:
      - description: Password update details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.PasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update password
      tags:
      - Account
  /auth/password/set:
    post:
      consumes:
      - application/json
      description: Sets a password for the first time on an existing account using
        invite verification
      parameters:
      - description: Password setup details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.SetPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Set password for new account
      tags:
      - Account
  /auth/verify/resend:
    post:
      consumes:
      - application/json
      description: Resends the email verification link to the user's email address
      parameters:
      - description: Cloudflare token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.ResendVerifyEmailPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Resend verification email
      tags:
      - Account
...
swagger: "2.0"